+++ internal/infra/websocket/client.go
